Move gen tangents into function
