+++ src/Packager/Packager.h
@@ -4,4 +4,5 @@
 namespace rosy_packager
 {
     void optimize_mesh(const std::shared_ptr<rosy_logger::log>& l, rosy_asset::mesh& asset_mesh);
+    rosy::result generate_tangents(const std::shared_ptr<rosy_logger::log>& l, rosy_asset::asset& asset);
 }
